Keep database connection alive
A couple of requests per day are failing with the following error:
  sqlalchemy.exc.OperationalError:
  (psycopg2.OperationalError) terminating connection due to
  administrator command
Because there are only a couple of requests a day, about 30 % or so are
failing because of this.

It's probably caused by the database being hobby tier and the
connections being occasionally killed because of it. Hopefully using the
database every minute should cause the problems to be noticed more
often in this task rather than in requests from GitHub.